[#35] : Added working example of full GPU collision detection

+++ src/manta_math_single.cpp
@@ -82,15 +82,15 @@ math::real math::getScalar(const math::Vector &v) {
 }
 
 math::Vector4 math::getVector4(const math::Vector &v) {
-	return Vector4(v.x, v.y, v.z, v.w);
+	return Vector4((float)v.x, (float)v.y, (float)v.z, (float)v.w);
 }
 
 math::Vector3 math::getVector3(const math::Vector &v) {
-	return Vector3(v.x, v.y, v.z);
+	return Vector3((float)v.x, (float)v.y, (float)v.z);
 }
 
 math::Vector2 math::getVector2(const math::Vector &v) {
-	return Vector2(v.x, v.y);
+	return Vector2((float)v.x, (float)v.y);
 }
 
 math::Generic math::add(const math::Generic &v1, const math::Generic &v2) {
